big fixes and UserDto creation

+++ src/main/java/com/infoschool/infoschool/service/UserService.java
@@ -5,11 +5,16 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.infoschool.infoschool.dto.request.SignupRequest;
+import com.infoschool.infoschool.dto.request.UserDto;
 import com.infoschool.infoschool.dto.request.UserDtoForm;
 import com.infoschool.infoschool.dto.request.UserRegistrarionToCourseDto;
 import com.infoschool.infoschool.model.Course;
+import com.infoschool.infoschool.model.Role;
 import com.infoschool.infoschool.model.User;
+import com.infoschool.infoschool.repository.RoleRepository;
 import com.infoschool.infoschool.repository.UserRepository;
+import com.infoschool.infoschool.util.ERole;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -21,15 +26,18 @@ public class UserService {
     private final UserRepository userRepository;
     @Autowired
     private final CourseService courseService;
+    @Autowired
+    private final RoleRepository roleRepository;
 
-    public UserService(UserRepository userRepository, CourseService courseService) {
+    public UserService(UserRepository userRepository, CourseService courseService, RoleRepository roleRepository) {
+        this.roleRepository = roleRepository;
         this.userRepository = userRepository;
         this.courseService = courseService;
     }
 
 
 
-    public User addUser(UserDtoForm user) {
+    public User registerUser(UserDtoForm user) {
         try {
             log.info("Adding new user: {}", user);
             User newUser = new User();
@@ -40,6 +48,49 @@ public class UserService {
             newUser.setAddress(user.getAddress());
             newUser.setBirthDate(user.getBirthDate());
             newUser.setRole(user.getRole());
+            return userRepository.save(newUser);
+        } catch (Exception e) {
+            log.error("Error adding user: {}", user, e);
+            return null;
+        }
+    }
+
+    public User addUser(SignupRequest user) {
+        try {
+            
+            log.info("Adding new user: {}", user);
+            User newUser = new User();
+            newUser.setName(user.getName());
+            newUser.setSurname(user.getSurname());
+            newUser.setEmail(user.getEmail());
+            newUser.setPassword(user.getPassword());
+            newUser.setAddress(user.getAddress());
+            newUser.setBirthDate(user.getBirthDate());
+            String strRole = user.getRole();
+            Role role = new Role();
+
+            if (strRole == null) {
+                role = roleRepository.findByName(ERole.ROLE_USER)
+                        .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+            } else {
+                    switch (strRole.toLowerCase()) {
+                        case "admin":
+                            Role adminRole = roleRepository.findByName(ERole.ROLE_ADMIN)
+                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+                            role = adminRole;
+                            break;
+                        case "teacher":
+                            role = roleRepository.findByName(ERole.ROLE_TEACHER)
+                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+                            break;
+                        default:
+                            Role userRole = roleRepository.findByName(ERole.ROLE_USER)
+                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+                            role = userRole;
+                    }
+            }
+
+            newUser.setRole(role);
             return userRepository.save(newUser);
         } catch (Exception e) {
             log.error("Error adding user: {}", user, e);
@@ -69,9 +120,19 @@ public class UserService {
     public void deleteUserById(Long id) {
         try {
             log.info("Deleting user by ID: {}", id);
-            userRepository.deleteById(id);
+            Optional<User> user = userRepository.findById(id);
+            if (user.isPresent()) {
+                userRepository.deleteById(id);
+            } else {
+                log.warn("User not found for deletion: {}", id);
+                throw new RuntimeException("User not found");
+            }
+        } catch (RuntimeException e) {
+            log.error("Error deleting user by ID: {}", id, e);
+            throw new RuntimeException("Error deleting user by ID: " + e.getMessage(), e);
         } catch (Exception e) {
             log.error("Error deleting user by ID: {}", id, e);
+            throw new RuntimeException("Error deleting user by ID: " + e.getMessage(), e);
         }
     }
 
@@ -85,18 +146,23 @@ public class UserService {
         }
     }
 
-    public User edit(UserDtoForm user) {
+    public User edit(UserDto user) {
         try {
             log.info("Editing user: {}", user);
-            User existingUser = userRepository.findById(user.getUserId()).orElse(null);
+            User existingUser = userRepository.findById(user.getId()).orElse(null);
             if (existingUser != null) {
                 existingUser.setName(user.getName());
                 existingUser.setSurname(user.getSurname());
                 existingUser.setEmail(user.getEmail());
-                existingUser.setPassword(user.getPassword());
                 existingUser.setAddress(user.getAddress());
                 existingUser.setBirthDate(user.getBirthDate());
-                existingUser.setRole(user.getRole());
+                Role role = roleRepository.findById(user.getRoleId()).orElse(null);
+                if (role != null) {
+                    existingUser.setRole(role);
+                } else {
+                    log.warn("Role not found for user: {}", user.getRoleId());
+                    throw new RuntimeException("Role not found");
+                }
                 return userRepository.save(existingUser);
             } else {
                 log.warn("User not found for editing: {}", user);
